Update navigation, move summaries under substances
