+++ pages/index/substances/deliriants/index.vue
@@ -170,9 +170,9 @@
 
     <h3> See Also </h3>
     <ul>
-      <li> <nuxt-link to="/summaries/"> Substance Summary Index </nuxt-link> </li>
-      <li> <nuxt-link to="/summaries/psychedelics/visual"> Visual Psychedelic Effects </nuxt-link> </li>
-      <li> <nuxt-link to="/summaries/dissociatives/"> Dissociative Subjective Effects </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/"> Substance Summary Index </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/psychedelics/visual"> Visual Psychedelic Effects </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/dissociatives/"> Dissociative Subjective Effects </nuxt-link> </li>
     </ul>
