+++ pages/index/substances/index.vue
@@ -1,61 +1,121 @@
 <template>
-  <div class="pageContent">
+  <div class="pageContent substanceArticles">
     <fa
-      :icon="['far', 'sitemap']"
+      :icon="['far', 'flask']"
       class="fa categoryIcon" />
-    <h1> Subjective Effect Index </h1>
-    <description />
-    substance list go here
-  </div>
+    <h1> Substance Summaries </h1>
+
+    <p>The following articles break down the specific subjective effects of hallucinogens
+    into simple and easy to understand descriptions with accompanying image replications.
+    This is done without relying on metaphor, analogy or personal trip reports.</p>
+
+    <p>The descriptions include comprehensive intensity scales which are applied to both
+    hallucinogenic substance classes and specific compounds. These intensity scales describe
+    the various experiential changes that can occur at specific ranges as the dosage is increased.</p>
+
+    <p>Alongside of these, we have also included more general and summarised descriptions of the
+    various effects which commonly occur under the influence of the three main classes of
+    hallucinogen: psychedelics, dissociatives, and deliriants. However, please keep in mind that
+    although the effect descriptions within these summaries are significantly shortened, their
+    full-length articles are linked above each effect.</p>
 </template>
 
 <script>
-import Tabs from "@/components/substances/index/Tabs";
-import Tab from "@/components/substances/index/Tab";
-import Category from "@/components/substances/index/Category";
-import Column from "@/components/substances/index/Column";
-import Actions from "@/components/substances/index/Actions";
-import Description from "@/components/substances/index/Description";
-import Blob from "@/components/substances/index/Blob";
+import Panel from '@/components/Panel';
+import Category from '@/components/Category';
 
 export default {
   components: {
-    Tabs,
-    Tab,
-    Category,
-    Actions,
-    Column,
-    Description,
-    Blob
+    Panel,
+    Category
   },
   head() {
     return {
-      title: "Subjective Effect Index"
+      title: "Summaries"
     };
   },
-  data() {
-    return {
-      activeTab: this.$route.query.type
-    };
-  },
+  scrollToTop: true,
   computed: {
-    substances() {
-      return this.$store.state.substances;
-    },
-  },
-  watchQuery: ['type'],
-  methods: {
-    filterSubstancesByTag(...tags) {
-      return this.substances.filter(substance =>
-        tags.every(tag => substance.tags.indexOf(tag) > -1)
-      );
-    },
-    selectTab(name) {
-      this.activeTab = name;
+    effects() {
+      return this.$store.state.effects;
     }
   },
-  async fetch ({ store }) {
-    await store.dispatch("getSubstances");
+  methods: {
+    filterEffectsByTag(...tags) {
+      return this.effects.filter(effect =>
+        tags.every(tag => effect.tags.indexOf(tag) > -1)
+      );
+    },
+    getEffectsInSpecificOrder(...names) {
+      let effects = [];
+      names.forEach((name) => {
+        let foundEffect = this.effects.find((effect) => effect.name.toLowerCase() === name.toLowerCase());
+        if (foundEffect) effects.push(foundEffect);
+        });
+      return effects;
+    }
   },
+  async fetch({ store }) {
+    await store.dispatch("getEffects");
+  }
 };
 </script>
+
+
+<style scoped>
+
+  h4 {
+    letter-spacing: 2px;
+    color: #444;
+    margin: 0.25em 0;
+    font-size: 13pt;
+    font-family: 'Titillium Web';
+  }
+
+  .summaryColumns {
+    display: flex;
+    flex-direction: row;
+    align-items: flex-start;
+    flex-wrap: wrap;
+  }
+
+  .categoriesContainer {
+    display: flex;
+    flex: 1;
+    flex-direction: column;
+    flex-wrap: wrap;
+    margin-right: 1em;
+  }
+
+  .summaryContentsTitle {
+    margin: 1em 0 0.25em 0;
+    color: #333;
+    font-size: 14pt;
+    letter-spacing: 0px;
+    text-transform: none;
+  }
+
+  .categoriesFullLink {
+    font-size: 12pt;
+    color: #666;
+    font-style: italic;
+    font-family: 'Titillium Web', -apple-system, BlinkMacSystemFont, "Segoe UI",
+    Roboto, "Helvetica Neue", Arial, sans-serif;
+  }
+
+  .effectList {
+    color: #AAA;
+    padding-left: 1em;
+  }
+
+  @media(max-width: 1100px) {
+    .columns {
+      display: block;
+    }
+  }
+
+
+
+
+
+</style>
