+++ pages/index/substances/psychedelics/cognitive.vue
@@ -114,9 +114,9 @@
 
     <h3> See Also </h3>
     <ul>
-      <li> <nuxt-link to="/summaries/"> Substance Summary Index </nuxt-link> </li>
-      <li> <nuxt-link to="/summaries/psychedelics/visual"> Visual Psychedelic Effects </nuxt-link> </li>
-      <li> <nuxt-link to="/summaries/psychedelics/miscellaneous"> Miscellaneous Psychedelic Effects </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/"> Substance Summary Index </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/psychedelics/visual"> Visual Psychedelic Effects </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/psychedelics/miscellaneous"> Miscellaneous Psychedelic Effects </nuxt-link> </li>
     </ul>
 
   </div>
