+++ pages/index/substances/psychedelics/miscellaneous.vue
@@ -116,9 +116,9 @@
 
     <h3> See Also </h3>
     <ul>
-      <li> <nuxt-link to="/summaries/"> Substance Summary Index </nuxt-link> </li>
-      <li> <nuxt-link to="/summaries/psychedelics/cognitive"> Cognitive Psychedelic Effects </nuxt-link> </li>
-      <li> <nuxt-link to="/summaries/psychedelics/visual"> Visual Psychedelic Effects </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/"> Substance Summary Index </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/psychedelics/cognitive"> Cognitive Psychedelic Effects </nuxt-link> </li>
+      <li> <nuxt-link to="/substances/psychedelics/visual"> Visual Psychedelic Effects </nuxt-link> </li>
     </ul>
 
   </div>
